requirements.txt
anthropic>=0.39.0
openai>=1.40.0
streamlit>=1.40.0
python-dotenv>=1.0.0
python-docx>=1.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
